+++ src/browser/ColorContrastCache.ts
@@ -4,7 +4,7 @@
  */
 
 import { IColorContrastCache } from './Types';
-import { IColor } from '../common/Types';
+import { IColor } from '../common/primitives/Types';
 import { TwoKeyMap } from '../common/primitives/MultiKeyMap';
 
 export class ColorContrastCache implements IColorContrastCache {
