+++ src/browser/CoreBrowserTerminal.ts
@@ -46,7 +46,8 @@ import { KeyboardService } from './services/KeyboardService';
 import { channels, color, rgb } from '../common/primitives/Color';
 import { CoreTerminal } from '../common/CoreTerminal';
 import * as Browser from '../common/primitives/Platform';
-import { ColorRequestType, IColorEvent, ITerminalOptions, KeyboardResultType, SpecialColorIndex } from '../common/Types';
+import { ColorRequestType, IColorEvent, KeyboardResultType, SpecialColorIndex } from '../common/primitives/Types';
+import { ITerminalOptions } from '../common/Types';
 import { DEFAULT_ATTR_DATA } from '../common/primitives/buffer/BufferLine';
 import { IBuffer } from '../common/primitives/buffer/Types';
 import { C0, C1ESCAPED } from '../common/primitives/data/EscapeSequences';
