+++ src/browser/OscLinkProvider.test.ts
@@ -8,7 +8,7 @@ import { OscLinkProvider } from './OscLinkProvider';
 import { ILink } from './Types';
 import { createCellData, MockBufferService, MockOptionsService } from '../common/TestUtils.test';
 import { IBufferService, IOscLinkService } from '../common/services/Services';
-import { IOscLinkData } from '../common/Types';
+import { IOscLinkData } from '../common/primitives/Types';
 import { IBufferLine } from '../common/primitives/buffer/Types';
 
 class TestOscLinkService implements IOscLinkService {
