+++ src/browser/renderer/shared/Types.ts
@@ -5,7 +5,7 @@
 
 import { Terminal } from '@xterm/xterm';
 import { ITerminal } from '../../Types';
-import { IDisposable } from '../../../common/Types';
+import { IDisposable } from '../../../common/primitives/Types';
 import type { IEvent } from '../../../common/primitives/Event';
 
 export interface IDimensions {
