+++ src/browser/services/LinkProviderService.ts
@@ -1,6 +1,6 @@
 import { ILinkProvider, ILinkProviderService } from './Services';
 import { Disposable, toDisposable } from '../../common/primitives/Lifecycle';
-import { IDisposable } from '../../common/Types';
+import { IDisposable } from '../../common/primitives/Types';
 
 export class LinkProviderService extends Disposable implements ILinkProviderService {
   declare public serviceBrand: undefined;
