+++ src/common/CoreTerminal.ts
@@ -26,7 +26,8 @@ import { InstantiationService } from './services/InstantiationService';
 import { LogService } from './services/LogService';
 import { BufferService, BufferServiceConstants } from './services/BufferService';
 import { OptionsService } from './services/OptionsService';
-import { IDisposable, IScrollEvent, ITerminalOptions, IParams } from './Types';
+import { IDisposable, IParams } from './primitives/Types';
+import { IScrollEvent, ITerminalOptions } from './Types';
 import { IAttributeData, IBufferSet } from './primitives/buffer/Types';
 import { CoreService } from './services/CoreService';
 import { MouseStateService } from './services/MouseStateService';
