+++ src/common/primitives/input/Keyboard.test.ts
@@ -1,7 +1,7 @@
 
 import { assert } from 'chai';
 import { evaluateKeyboardEvent } from './Keyboard';
-import { IKeyboardResult, IKeyboardEvent } from '../../Types';
+import { IKeyboardResult, IKeyboardEvent } from '../Types';
 
 /**
  * A helper function for testing which allows passing in a partial event and defaults will be filled
