+++ src/common/primitives/parser/ApcParser.ts
@@ -6,7 +6,7 @@
 import { IApcHandler, IHandlerCollection, ApcFallbackHandlerType, IApcParser, ISubParserStackState } from './Types';
 import { ParserConstants } from './Constants';
 import { utf32ToString } from '../input/TextDecoder';
-import { IDisposable } from '../../Types';
+import { IDisposable } from '../Types';
 import { LimitedStringBuilder } from '../StringBuilder';
 
 const EMPTY_HANDLERS: IApcHandler[] = [];
