+++ src/common/primitives/parser/OscParser.ts
@@ -6,7 +6,7 @@
 import { IOscHandler, IHandlerCollection, OscFallbackHandlerType, IOscParser, ISubParserStackState } from './Types';
 import { OscState, ParserConstants } from './Constants';
 import { utf32ToString } from '../input/TextDecoder';
-import { IDisposable } from '../../Types';
+import { IDisposable } from '../Types';
 import { LimitedStringBuilder } from '../StringBuilder';
 
 const EMPTY_HANDLERS: IOscHandler[] = [];
