+++ src/common/services/OptionsService.test.ts
@@ -5,7 +5,7 @@
 
 import { assert } from 'chai';
 import { OptionsService, DEFAULT_OPTIONS } from './OptionsService';
-import { IDisposable } from '../Types';
+import { IDisposable } from '../primitives/Types';
 
 describe('OptionsService', () => {
   describe('constructor', () => {
