+++ src/common/services/OptionsService.ts
@@ -5,7 +5,7 @@
 
 import { Disposable, toDisposable } from '../primitives/Lifecycle';
 import { isMac } from '../primitives/Platform';
-import { CursorStyle, IDisposable } from '../Types';
+import { CursorStyle, IDisposable } from '../primitives/Types';
 import { FontWeight, IOptionsService, ITerminalOptions } from './Services';
 import { Emitter } from '../primitives/Event';
