+++ src/common/services/OscLinkService.ts
@@ -3,7 +3,7 @@
  * @license MIT
  */
 import { IBufferService, IOscLinkService } from './Services';
-import { IOscLinkData } from '../Types';
+import { IOscLinkData } from '../primitives/Types';
 import { IMarker } from '../primitives/buffer/Types';
 
 export class OscLinkService implements IOscLinkService {
